nvme: fix signed/unsigned error with cuse ioctls
cuse defines the ioctl handler to take an int for
the command, but the kernel uses unsigned int.  So
cast the int parameter to unsigned int when doing
the switch.

Fixes issue #1613.

Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/4716
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ lib/nvme/nvme_cuse.c
@@ -555,7 +555,7 @@ cuse_ctrlr_ioctl(fuse_req_t req, int cmd, void *arg,
 		return;
 	}
 
-	switch (cmd) {
+	switch ((unsigned int)cmd) {
 	case NVME_IOCTL_ADMIN_CMD:
 		cuse_nvme_admin_cmd(req, cmd, arg, fi, flags, in_buf, in_bufsz, out_bufsz);
 		break;
@@ -580,7 +580,7 @@ cuse_ns_ioctl(fuse_req_t req, int cmd, void *arg,
 		return;
 	}
 
-	switch (cmd) {
+	switch ((unsigned int)cmd) {
 	case NVME_IOCTL_ADMIN_CMD:
 		cuse_nvme_admin_cmd(req, cmd, arg, fi, flags, in_buf, in_bufsz, out_bufsz);
 		break;
